requirements.txt
customtkinter>=5.2.0
pandas>=1.5.0
numpy>=1.23.0
Pillow>=9.0.0
chardet>=5.0.0
matplotlib>=3.5.0
openpyxl>=3.1.0
